% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.icweib.R
\name{print.icweib}
\alias{print.icweib}
\title{Print icweib object}
\usage{
\method{print}{icweib}(x, digits = 3, ...)
}
\arguments{
\item{x}{output returned by icweib function.}

\item{digits}{digits to print.}

\item{...}{other arguments to be passed from print function.}
}
\description{
This function prints the summary of the fitting results from icweib.
}
\examples{
data(tooth24)
fit <- icweib(L = left, R = right, data = tooth24, strata = dmf, covariates = ~sex)
fit

}
